Constructor overloaded
Used ChatGPT to learn how constructor overloading works in C++. In comparison to Java, you don't need four explicit constructors to do this, and there isn't a super() function. Instead, you can set a default value for the parameter, and if no value is given it will just use the default.